added security to views to be available ony for admin users

+++ packages/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Package
 
 from .forms import PackageForm
@@ -30,7 +31,12 @@ def package_detail(request, package_id):
     return render(request, 'packages/package_detail.html', context)
 
 
+@login_required
 def add_package(request):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only gym administrator can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = PackageForm(request.POST, request.FILES)
         if form.is_valid():
@@ -50,7 +56,12 @@ def add_package(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_package(request, package_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only gym administrator can do that.')
+        return redirect(reverse('home'))
+
     package = get_object_or_404(Package, pk=package_id)
     if request.method == 'POST':
         form = PackageForm(request.POST, request.FILES, instance=package)
@@ -72,7 +83,12 @@ def edit_package(request, package_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_package(request, package_id):
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only gym administrator can do that.')
+        return redirect(reverse('home'))
+
     package = get_object_or_404(Package, pk=package_id)
     package.delete()
     messages.success(request, 'Successfully deleted package')
